Enum disqrust::JobStatus

pub enum JobStatus {
    FastAck,
    AckJob,
    NAck,
}

Once a job execution finishes, change its status by performing one of this actions.

Variants

FastAck

Performs a best effort cluster wide deletion of the job.

AckJob

Acknowledges the execution of the jobs.

NAck

Puts back the job in the queue ASAP.